% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{allpair_RandIndex}
\alias{allpair_RandIndex}
\title{Use RandIndex to evaluate number of clusters}
\usage{
allpair_RandIndex(results)
}
\arguments{
\item{results}{List with results from trajectories() function}
}
\value{
A data frame with RandIndex for all pairs from trajectories results.
Note all pairs means lower triangle plus diagonal of an all-pairs symmetric
matrix.
}
\description{
Use RandIndex to evaluate number of clusters
}
