% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{bench_cores}
\alias{bench_cores}
\title{Function to benchmark various core combinations for components of clustra}
\usage{
bench_cores(FUN, ..., max2 = 5, reps = 4)
}
\arguments{
\item{FUN}{Function to benchmark.}

\item{...}{Additional arguments passed to `FUN`.}

\item{max2}{Largest power of 2 in cores (e.g. max2 = 3 implies 1, 2, 4, and 8
cores will be tested. Core control is via openblasctl package)}

\item{reps}{Number of replicates for each core setting}
}
\description{
Function to benchmark various core combinations for components of clustra
}
