% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{check_df}
\alias{check_df}
\title{Function to check if non-zero group members have enough data for spline fit
degrees of freedom.}
\usage{
check_df(group, loss, data, fp)
}
\arguments{
\item{group}{Group membership vector for each id.}

\item{loss}{A matrix with rows of computed loss values of each id across all models}

\item{data}{A data.table with data. See \code{link{trajectories}}.}

\item{fp}{Fitting parameters. See \code{link{trajectories}}.}
}
\value{
Returns the vector of group membership of id's either unchanged or changed
to have sufficient data in non-zero groups.
}
\description{
Function to check if non-zero group members have enough data for spline fit
degrees of freedom.
}
