% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{clustra}
\alias{clustra}
\title{Cluster trajectories}
\usage{
clustra(data, k, starts = list(ns = 5, nid = 20), cores = list(e_mc =
  1, m_mc = 1, bam_nthreads = 1, blas = 1), maxdf = 50,
  verbose = FALSE)
}
\arguments{
\item{data}{Data frame with response measurements, one per observation.
Column names are id, time, response, group.}

\item{k}{Number of clusters (groups)}

\item{starts}{A list with two components: ns is the number of sampled starts
to evaluate as starting cluster assignments, nid is the number of
trajectories to sample for the starts.}

\item{cores}{A list control structure to manage cores used for parallel 
processing of various parts.}

\item{maxdf}{Maximum spline order to use in tps fits.}
}
\description{
Cluster trajectories
}
