% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clustra_env}
\alias{clustra_env}
\title{Get or set \code{\link{clustra}} environment variables.}
\usage{
clustra_env(...)
}
\arguments{
\item{...}{Any number of string constants that ask for or assign various
parameters in \code{.clustra_env} environment. For example, \code{"cor$e_mc"}
will return the current \code{cor$e_mc} parameter and \code{"cor$e_mc = 8"}
will change it to 8.}
}
\description{
Several parameters can be set with one call but only one parameter can
be returned. Consequently all parameters after the first returned 
parameter are ignored. Note that limited checking for syntax
is performed so be careful!
}
\details{
An empty parameter list will print all parameters in the environment.
}
\seealso{
\code{\link{.clustra_env_gen}}, \code{\link{.clustra_env_cor}},
 \code{\link{.clustra_env_clu}}, \code{\link{.clustra_env_ran}}.
}
