% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{clustra_par}
\alias{clustra_par}
\title{Function to set and maintain clustra parameters and working directory.}
\usage{
clustra_par(playdir, parname = "clustra_par.json", PL = NULL)
}
\arguments{
\item{playdir}{Character string directory where filename is expected and written.}

\item{parname}{Character string filename.}

\item{PL}{A list of lists parameter data structure.}
}
\description{
Function to set and maintain clustra parameters and working directory.
}
\details{
Objective is to simplify clustra function invocation and manage sets of
clustra runs using a \code{jsonlite} file. Sets working directory to playdir.
Parameters PL set in the following order: if PL is supplied, it is written to
a json file and returned; if PL is NULL, attempts to find parname file in
playdir and use that; if PL is NULL and parname file is not present, PL is
set to defaults.
}
