% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{clustra_rand}
\alias{clustra_rand}
\title{clustra_rand:
Performs \code{\link{clustra}} runs for several k and presents a k selection
index.}
\usage{
clustra_rand(
  data,
  k,
  cores,
  replicates = 10,
  fp = list(maxdf = 30, iter = 10),
  save = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{The data (see \code{\link{clustra}} description).}

\item{k}{Vector of k values to try.}

\item{cores}{A vector of core assignments for multicore components. See 
\code{\link{trajectories}}.}

\item{replicates}{Number of replicates for each k.}

\item{fp}{Fitting parameters. See \code{link{trajectories}}.}

\item{save}{Logical. When TRUE, save all results as file \code{results.Rdata}.}

\item{verbose}{Logical. When TRUE, information about each run of clustra is printed.}
}
\value{
See \code{\link{allpair_RandIndex}}
}
\description{
Available selection indices are "silhouette" and "rand".
Silhouette computes a proxy silhouette index based on distances to cluster
centers rather than trajectory pairs. The cost is essentially that of
running clustra for several k as this information is
available directly from clustra.
}
\details{
Rand runs several random start replicates per k, resulting in a much slower
algorithm. Then prepares a Rand index comparison between
all pairs of clusterings.
}
