% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{data.prep}
\alias{data.prep}
\title{data.prep}
\usage{
data.prep(data, model)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{model}{See \code{\link{clustra}} description.}
}
\value{
Returns a data.table with variables id, response, and time based on the
model specification.
}
\description{
Internal function to check data and extract a minimal data.table with
standardized variable names according to the model.
}
