% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.clustra_env_gen}
\alias{.clustra_env_gen}
\title{Function to initialize components of `gen` list in .clustra_env environment}
\usage{
.clustra_env_gen(
  seed = 90,
  n_id = 1000,
  m_obs = 25,
  s_range = c(-50, -10),
  e_range = c(1095, 3650),
  plots = FALSE
)
}
\arguments{
\item{seed}{Seed to use in data generation.}

\item{n_id}{Number of id to generate.}

\item{m_obs}{Mean number of observation per `id` using}

\item{s_range}{A vector of length 2, giving the min and max limits of
uniformly generated start observation time.}

\item{e_range}{A vector of length 2, giving the min and max limits of
uniformly generated end observation time.}

\item{plots}{Logical to indicate if a sample of `id`s trajectories are
plotted.}
}
\description{
Function to initialize components of `gen` list in .clustra_env environment
}
