% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.clustra_env_ran}
\alias{.clustra_env_ran}
\title{Function to initialize components of `ran` list in \code{.clustra_env}
environment}
\usage{
.clustra_env_ran(ng_vec = c(2, 3, 4), replicates = 10)
}
\arguments{
\item{ng_vec}{Vector of `k` values to simulate.}

\item{replicates}{Number of replicates for each `k`.}
}
\description{
Function to initialize components of `ran` list in \code{.clustra_env}
environment
}
