% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.clustra_init}
\alias{.clustra_init}
\title{Function to create and initialize \code{.clustra_env}}
\usage{
.clustra_init(envir = .GlobalEnv)
}
\arguments{
\item{envir}{Environment where to create \code{.clustra_env}}
}
\description{
Function to create and initialize \code{.clustra_env}
}
