% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clustra_cor}
\alias{clustra_cor}
\alias{clustra_clu}
\title{Components of `cor` list in .clustra_env environment}
\usage{
clustra_cor(e_mc = 1, m_mc = 1, bam_nthreads = 1, blas = 1)

clustra_clu(seed = 79, maxdf = 50, iter = 8, starts = 4, idperstart = 20)
}
\arguments{
\item{e_mc}{Cores to use by mclapply of e_mc (expectation over k)}

\item{m_mc}{Cores to use by mclapply of m_mc (maximization across k)}

\item{bam_nthreads}{Threads to use by mgcv::bam()}

\item{blas}{Cores to use by OpenBLAS}

\item{seed}{Cores to use by mclapply of e_mc (expectation over k)}

\item{maxdf}{Basis dimension of smooth term (mgcv::s(), parameter k).}

\item{iter}{Maximum number of iterations.}

\item{starts}{Number of random starts.}

\item{idperstart}{Number of `id`s to sample for random starts.}
}
\description{
Components of `cor` list in .clustra_env environment

Components of `clu` list in .clustra_env environment
}
\seealso{
[mgcv::s()]
}
