% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_long_data}
\alias{gen_long_data}
\title{Data Generators}
\usage{
gen_long_data(n_id, m_obs, s_range = c(0, 0), e_range = c(50, 100),
  center = 100, noise = c(0, center/20), n_typ = 3, plots = 0)
}
\arguments{
\item{n_id}{Number of trajectories}

\item{m_obs}{Poisson mean for number of observations in a trajectory}

\item{s_range}{Vector of length 2 giving last observation time range, 
c(min, max). Lengths are uniformly distributed over this range.}

\item{e_range}{Vector of length 2 giving last observation time range, 
c(min, max). Lengths are uniformly distributed over this range.}

\item{noise}{Vector of length 2 giving the mean and standard deviation, 
c(mean, sd), of added N(mean, sd) noise.}

\item{plots}{Number of plots of randomly selected trajectories to check
generated data}
}
\description{
Collection of longitudinal responses with possibly varying lengths and 
varying numbers of observations.
}
\section{Details}{

Generate longitudinal data for a response variable. Trajectories start
at time uniformly distributed in s_range and end at time uniformly 
distributed in e_range. Number of observations in a trajectory is 
Poisson(m_obs). The result is a number of trajectories, all starting at 
time 0, with different time spans, and with independently different numbers 
of observations within the time spans. Each trajectory follows a randomly 
selected response function with added N(mean, sd) error.
}

\section{Value}{

A data frame with one response per row and three columns: id (an integer in 
1:n_id), time (an iteger in 0 to last observation time), and response.
}

