% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gendata}
\alias{gendata}
\title{gendata}
\usage{
gendata(n_id, m_obs, s_range, e_range, min_obs, reference, noise)
}
\arguments{
\item{n_id}{See parameters of {\code{\link{gen_traj_data}}}.}

\item{m_obs}{See parameters of {\code{\link{gen_traj_data}}}.}

\item{s_range}{See parameters of {\code{\link{gen_traj_data}}}.}

\item{e_range}{See parameters of {\code{\link{gen_traj_data}}}.}

\item{min_obs}{See parameters of {\code{\link{gen_traj_data}}}.}

\item{reference}{See parameters of {\code{\link{gen_traj_data}}}.}

\item{noise}{See parameters of {\code{\link{gen_traj_data}}}.}
}
\value{
A list of length \code{sum(n_id)}, where each element is a matrix output by {\code{\link{oneid}}}.
}
\description{
Generates data for up to three trajectory clusters
}
\details{
Time support of each \code{id} is at least \verb{s . . . 0 . . . . e}, where \code{s} is in
\code{s_range} and \code{e} is in \code{e_range}.
}
