% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{ic_fun}
\alias{ic_fun}
\title{Function to test information criteria. Not exported and used by internal
function \code{kchoose}.}
\usage{
ic_fun(cl, data, fn)
}
\arguments{
\item{cl}{Output from \code{clustra} function.}

\item{data}{A valid data set for \code{clustra}.}

\item{fn}{Character file name for output.}
}
\value{
Numerical value of computed AIC. Also writes a line of computed information
criteria to \code{fn} file for each k.
}
\description{
Function to test information criteria. Not exported and used by internal
function \code{kchoose}.
}
