% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{kchoose}
\alias{kchoose}
\title{A test function to evaluate information criteria for several k values. Not
exported and only for debugging internal use.}
\usage{
kchoose(K, var = 5, maxdf = 10, mc = 1, fn = "ic.txt")
}
\arguments{
\item{K}{Integer vector of k values to try.}

\item{var}{A numerical value of noise variance in generated data.}

\item{maxdf}{Fitting parameters. See \code{\link{trajectories}}.}

\item{mc}{Number of cores to use. Increase up to largest k, or number of cores
available, whichever is less. (On hyperthreaded cores, up to 2x number of
cores.)}

\item{fn}{Character file name for output.}
}
\description{
A test function to evaluate information criteria for several k values. Not
exported and only for debugging internal use.
}
