% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{mse_g}
\alias{mse_g}
\alias{mxe_g}
\title{Loss functions}
\usage{
mse_g(g, pred, data)

mxe_g(g, pred, data)
}
\arguments{
\item{g}{Integer group number.}

\item{pred}{List. Element g is output of \code{\link{pred_g}} for group g.}

\item{data}{See \code{\link{trajectories}}}
}
\description{
\code{mse_g} Computes mean-squared loss for each group.
\code{mxe_g} is maximum loss.
}
