% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{oneid}
\alias{oneid}
\title{Generates data for one \code{id}}
\usage{
oneid(vars, clusters, id, n_obs, type, start, end, smin, emax)
}
\arguments{
\item{id}{A unique integer.}

\item{n_obs}{An integer number of observations to produce.}

\item{type}{Response type, 1 is constant, 2 is a sin curve portion, and 3 is a sigmoid
portion.}

\item{start}{Negative integer giving time of first observation.}

\item{end}{Positive integer giving time of last observation.}

\item{smin}{The smallest possible \code{start} value among all \code{id}s. Currently not used.}

\item{emax}{The largest possible \code{end} value among all \code{id}s. Used to scale sin and
sigmoid support.}
}
\value{
An \code{n_obs} by \code{3 + length(vars)} matrix with columns \code{id}, \code{true_group},
\code{time}, and responses \code{vars}.
}
\description{
Generates data for one \code{id}
}
