% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_silhouette}
\alias{plot_silhouette}
\title{Plots a list item, a silhouette, from the result of \code{clustra_sil} along
with the average silhouette value.}
\usage{
plot_silhouette(sil)
}
\arguments{
\item{sil}{A data frame that is a list item returned by \code{clustra_sil}.}
}
\value{
Returns invisibly the average silhouette value.
}
\description{
Plots a list item, a silhouette, from the result of \code{clustra_sil} along
with the average silhouette value.
}
