% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{plot_tps}
\alias{plot_tps}
\title{Various functions to evaluate results of clustering, like RandIndex
computation and plots of spline fits.}
\usage{
plot_tps(dat, data = TRUE)
}
\arguments{
\item{dat}{}
}
\description{
Plots spline fit (modify to use acutal bam results), possibly with data
}
\section{Details}{

}

