% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{rand_clustra}
\alias{rand_clustra}
\title{Performs clustra runs for several k and several random start replicates per k.
Then prepares a Rand index comparison between all pairs of clusterings, that
are displayed in a matrix plot.}
\usage{
rand_clustra(data, k, PL, save = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{The data (see clustra description).}

\item{k}{Vector of k values to run.}

\item{PL}{A list data structure (a list of lists) giving all needed
parameters for the clustra runs.}

\item{save}{Logical. When TRUE, save all results as file results.Rdata}

\item{verbose}{Logical. When TRUE, information about each run of clustra is
printed.}
}
\description{
Performs clustra runs for several k and several random start replicates per k.
Then prepares a Rand index comparison between all pairs of clusterings, that
are displayed in a matrix plot.
}
