% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{start_groups}
\alias{start_groups}
\title{Function to assign starting groups.}
\usage{
start_groups(
  data,
  k,
  nstart = clustra_env("clu$starts"),
  nid = clustra_env("clu$idperstart"),
  maxdf = clustra_env("clu$maxdf"),
  cores = clustra_env("cor"),
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame with response measurements, one per observation.
Column names are id, time, response, group. Note that id are already
sequential starting from 1. This affects expanding group numbers to ids.}

\item{k}{Number of clusters (groups).
"bestrand" is implemented.}

\item{nstart}{(see clu component of .clustra_env)}

\item{nid}{(see clu component of .clustra_env)}

\item{maxdf}{(see clu component of .clustra_env)}

\item{cores}{(see cor component of .clustra_env)}

\item{verbose}{Turn on more output for debugging.}
}
\description{
Function to assign starting groups.
}
