% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{tps_g}
\alias{tps_g}
\title{Function to fit thin plate spline (tps) to a group with
\code{\link[mgcv]{bam}} from the \code{mgcv} package. Crossvalidation does
not know about zero weights, resulting in different smoothing parameters, so
subset parameter (rather than zero weights) is used to ensure correct
crossvalidation sampling.}
\usage{
tps_g(g, data, maxdf, nthreads)
}
\arguments{
\item{g}{Integer group number.}

\item{data}{See \code{\link{clustra}} description.}

\item{maxdf}{See \code{\link{trajectories}} description.}

\item{nthreads}{Controls \code{\link[mgcv]{bam}} threads.}
}
\description{
Function to fit thin plate spline (tps) to a group with
\code{\link[mgcv]{bam}} from the \code{mgcv} package. Crossvalidation does
not know about zero weights, resulting in different smoothing parameters, so
subset parameter (rather than zero weights) is used to ensure correct
crossvalidation sampling.
}
