% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{traj_rep}
\alias{traj_rep}
\title{Function to run trajectories inside mclapply with one core.}
\usage{
traj_rep(group, data, k, maxdf, conv)
}
\arguments{
\item{group}{Vector of starting group values for unique id's.}

\item{data}{The data (see \code{\link{clustra}} description).}

\item{k}{Integer number of clusters.}

\item{maxdf}{Fitting parameters. See \code{\link{trajectories}}.}

\item{conv}{Fitting parameters. See \code{\link{trajectories}}.}
}
\value{
See return of \code{\link{trajectories}}.
}
\description{
Function to run trajectories inside mclapply with one core.
}
