% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{xit_report}
\alias{xit_report}
\title{xit_report examines trajectories output to name what was concluded, such as
convergence, maximum iterations reached, a zero cluster, etc. Multiple
conclusions are possible as not all are mutually exclusive.}
\usage{
xit_report(cl, maxdf, iter)
}
\arguments{
\item{cl}{Output structure from trajectories() function}

\item{maxdf}{Fitting parameters. See \code{link{trajectories}}.}

\item{iter}{Fitting parameters. See \code{link{trajectories}}.}
}
\description{
xit_report examines trajectories output to name what was concluded, such as
convergence, maximum iterations reached, a zero cluster, etc. Multiple
conclusions are possible as not all are mutually exclusive.
}
