% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{allpair_RandIndex}
\alias{allpair_RandIndex}
\title{Use RandIndex to evaluate number of clusters}
\usage{
allpair_RandIndex(results)
}
\arguments{
\item{results}{List with results from trajectories() function}
}
\description{
Use RandIndex to evaluate number of clusters
}
