% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{allpair_RandIndex}
\alias{allpair_RandIndex}
\title{allpair_RandIndex is used internally by \code{\link{clustra_rand}} and
provides its return value.}
\usage{
allpair_RandIndex(results)
}
\arguments{
\item{results}{A list with each element packed internally by the
\code{\link{clustra_rand}} function.}
}
\value{
A data frame with \code{\link[MixSim]{RandIndex}} for all pairs from
trajectories results. The data frame names and its format is intended to be
the input for \code{\link{rand_plot}}. Note that all pairs is the lower
triangle plus diagonal of an all-pairs symmetric matrix.
}
\description{
Runs \code{\link[MixSim]{RandIndex}} for all pairs of cluster results in its
list input and produces a matrix for use by \code{\link{rand_plot}}.
}
