% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustra-package.R
\docType{package}
\name{clustra-package}
\alias{clustra-package}
\title{clustra-package}
\description{
Clusters trajectories (unequally spaced and unequal length time series) on
a common time axis. Clustering proceeds by an EM algorithm that iterates
switching between fitting a thin plate spline (TPS) to combined responses
within each cluster (M-step) and reassigning cluster membership based on the
nearest fitted TPS (E-step). Initial cluster assignments are random or
distant trajectories. The fitting is done with the \emph{mgcv} package function
\emph{bam}, which scales well to very large data sets. Additional parallelism
available via multicore on unix and mac platforms.
}
\author{
George Ostrouchov, David Gagnon, Hanna Gerlovin
}
\keyword{#}
\keyword{Import}
\keyword{Package}
\keyword{operators}
\keyword{package}
