% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustra-package.R
\docType{package}
\name{clustra-package}
\alias{clustra-package}
\title{clustra-package}
\description{
Clusters medical trajectories (unequally spaced and unequal lengths) aligned
by an intervention time. Performs k-means clustering, where each mean is a
thin plate spline fit to all points in a cluster. Distance is MSE across
trajectory points to cluster spline. Provides silhouette plots and Adjusted
Rand Index evaluations of the number of clusters. Scales well to large data
with multicore parallelism available to speed computation.
}
\author{
George Ostrouchov, David Gagnon, Hanna Gerlovin
}
\keyword{#}
\keyword{Import}
\keyword{Package}
\keyword{operators}
\keyword{package}
