% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustra-package.R
\docType{package}
\name{clustra-package}
\alias{clustra-package}
\title{clustra-package}
\description{
Performs k-means clustering, where each mean is a thin plate spline
fit to all points of trajectories in cluster.
}
\author{
George Ostrouchov, David Gagnon, Hanna Gerlovin
}
\keyword{Package}
