% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{clustra}
\alias{clustra}
\title{Cluster trajectories}
\usage{
clustra(data, k, PL, group = NULL, verbose = FALSE, plot = FALSE)
}
\arguments{
\item{data}{Data frame with response measurements, one per observation.
Column names are id, time, response, group.}

\item{k}{Number of clusters (groups).}

\item{PL}{A list of lists control structure created by \code{clustra_par}.}

\item{group}{A vector of initial cluster assignments for unique id's in data.
Normally, this is NULL and starts are provided by \code{start_groups()}.}

\item{verbose}{Logical to turn on more output during fit iterations.}

\item{plot}{Logical to indicate}
}
\description{
Most users will run the \code{clustra()} function, which takes care of
starting values and completes kmeans iteration according to parameters in
\code{PL} supplied by \code{clustra_par()} (see \code{demo} directory for
examples).
}
