% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{clustra}
\alias{clustra}
\title{Cluster trajectories}
\usage{
clustra(data, k, group = NULL, verbose = FALSE)
}
\arguments{
\item{data}{Data frame with response measurements, one response per observation.
Required variables are (id, time, response). Other variables are ignored.}

\item{k}{Number of clusters (groups).}

\item{group}{A vector of initial cluster assignments for unique id's in data.
Normally, this is NULL and good starts are provided by
\code{link{start_groups}}.}

\item{verbose}{Logical to turn on more output during fit iterations.}
}
\description{
Most users will run the \code{\link{clustra}} function, which takes care of
starting values and completes kmeans iteration according to parameters in
\code{.clustra_env} environment (See vignette("clustra_basic.Rmd") for
more details).
}
\details{
In addition to the shown parameters, detailed clustering and core
allocation parameters are in \code{.clustra_env} environment that can be
controlled with the \code{clustra_env} function.
}
