% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{clustra}
\alias{clustra}
\title{Cluster trajectories}
\usage{
clustra(
  data,
  k,
  group = NULL,
  fp = list(maxdf = 30, iter = 10, starts = 3, idperstart = 20, retry_max = 3),
  cores = c(e_mc = 1, m_mc = 1, nthreads = 1, blas = 1),
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame or, preferably, also a data.table with response measurements, one
response per observation. Required variables are (id, time, response).
Other variables are ignored.}

\item{k}{Number of clusters}

\item{group}{A vector of initial cluster assignments for unique id's in data.
Normally, this is NULL and good starts are provided by
\code{link{start_groups}}.}

\item{fp}{Fitting parameters. See \code{link{trajectories}}.}

\item{cores}{A vector of core assignments for multicore components. See
\code{link{trajectories}}.}

\item{verbose}{Logical to turn on more output during fit iterations.}
}
\description{
Most users will run the wrapper \code{\link{clustra}} function, which takes
care of starting values. See vignette("clustra_basic.Rmd") for
more details.
}
\examples{
set.seed(13)
data = gen_traj_data(n_id = 100, m_obs = 20, s_range = c(-50, -10),
              e_range = c(2*365, 5*365))
fp = list(maxdf = 20, iter = 5, starts = 2, idperstart = 20, retry_max = 3)
cl = clustra(data, k = 3, fp = fp, verbose = TRUE)
names(cl)
cl$tps

}
