% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clustra_env}
\alias{clustra_env}
\title{Get or set clustra environment variables.}
\usage{
clustra_env(...)
}
\arguments{
\item{...}{Any number of string constants that ask for or assign various
parameters in .clustra_env environment. For example, "cor$e_mc" will
return the cor$e_mc parameter and "cor$e_mc = 8" will assign 8 to that
parameter.}
}
\description{
Several parameters can be set with one call but only one parameter can
be returned. All parameters after the first returned parameter are ignored.
An empty parameter list will print all parameters in the environment.
Note that limited checking for syntax is performed so be careful!
}
\seealso{
[clustra_env_gen()], [clustra_env_cor()], [clustra_env_clu()], 
  [clustra_env_ran()].
}
