% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltime.R
\name{deltime}
\alias{deltime}
\title{Timing function}
\usage{
deltime(ltime = proc.time()["elapsed"], text = NULL)
}
\arguments{
\item{ltime}{Result of last call to deltime.}

\item{text}{Text to display along with elapsed time.}
}
\value{
"elapsed" component of current \code{\link{proc.time}}.
}
\description{
Timing function
}
