% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.clustra_env_clu}
\alias{.clustra_env_clu}
\title{Function to initialize components of `clu` list in .clustra_env environment}
\usage{
.clustra_env_clu(seed = 79, maxdf = 50, iter = 8, starts = 4, idperstart = 20)
}
\arguments{
\item{seed}{Cores to use by mclapply of e_mc (expectation over k)}

\item{maxdf}{Basis dimension of smooth term (mgcv::s(), parameter k).}

\item{iter}{Maximum number of iterations.}

\item{starts}{Number of random starts.}

\item{idperstart}{Number of `id`s to sample for random starts.}
}
\description{
Function to initialize components of `clu` list in .clustra_env environment
}
