% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_traj_data}
\alias{gen_traj_data}
\title{Data Generators}
\usage{
gen_traj_data(
  vars,
  clusters,
  m_obs,
  s_range,
  e_range,
  min_obs = 3,
  cv = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{vars}{A character vecor of variable names}

\item{clusters}{A matrix or a data frame with \code{length(vars)} rows and \code{2 + 2*length(vars)}
columns giving cluster size, noise coefficient of variation (defaults to 1),
variable means, and variable curve types (1 to 3), in that order.}

\item{m_obs}{Mean number of observation per id. Provides \code{lambda} parameter in
\code{\link[stats]{rpois}}.}

\item{s_range}{A vector of length 2, giving the min and max limits of uniformly generated
start observation time.}

\item{e_range}{A vector of length 2, giving the min and max limits of uniformly generated
end observation time.}

\item{min_obs}{Minimum number of observations in addition to zero time observation.}

\item{cv}{Coefficient of variation on generated data. If \code{clusters} is supplied as a
matrix of data.frame, it will override this parameter.}
}
\value{
A data table with one response per row and \code{3 + length(vars)}
columns: \code{id}, \code{true_group}, \code{time}, and a column for each of vars.
}
\description{
Generates a collection of longitudinal responses with possibly varying
lengths and varying numbers of observations. Support is
\code{start} . . . 0 . . . \code{end}, where
\code{start}~uniform(s_range) and \code{end}~uniform(e_range), so that
all trajectories are aligned at 0 but can start and end at different times.
Zero is the intervention time.
}
\section{Details}{

Generate longitudinal data for a response variable. Trajectories start
at time uniformly distributed in s_range and end at time uniformly
distributed in e_range. Number of observations in a trajectory is
Poisson(m_obs). The result is a number of trajectories, all starting at
time 0, with different time spans, and with independently different numbers
of observations within the time spans. Each trajectory follows a randomly
selected response function with added N(mean, sd) error.
}

\examples{
data = gen_traj_data(n_id = c(50, 100), m_obs = 20, s_range = c(-365, -14),
              e_range = c(0.5*365, 2*365))
head(data)
tail(data)

}
