% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_traj_data}
\alias{gen_traj_data}
\title{Data Generators}
\usage{
gen_traj_data(
  n_id,
  m_obs,
  s_range,
  e_range,
  reference = 100,
  noise = c(0, abs(reference/20)),
  k = 3,
  min_obs = 3
)
}
\arguments{
\item{n_id}{Number of id to generate.}

\item{m_obs}{Mean number of observation per id. Provides \code{lambda} parameter in
\code{\link[stats]{rpois}}.}

\item{s_range}{A vector of length 2, giving the min and max limits of uniformly generated
start observation time.}

\item{e_range}{A vector of length 2, giving the min and max limits of uniformly generated
end observation time.}

\item{reference}{A nominal response reference (for example, blood pressure is near 100, which
is the default)}

\item{noise}{Vector of length 2 giving the *mean* and *sd* of added N(mean, sd) noise.}

\item{k}{Number of response types (So far works only for k = 2 and 3)}

\item{min_obs}{Minimum number of observations in addition to zero time observation.}
}
\value{
A data frame with one response per row and three columns:
  \code{id}, \code{time}, \code{response}, and \code{group}.
}
\description{
Generates a collection of longitudinal responses with possibly varying
lengths and varying numbers of observations. Support is
[start . . . 0 . . . end], where
start~uniform(s_range) and end~uniform(e_range), so that
all trajectories are aligned at 0 but can start and end at different times.
Zero is the intervention time.
}
\section{Details}{

Generate longitudinal data for a response variable. Trajectories start
at time uniformly distributed in s_range and end at time uniformly 
distributed in e_range. Number of observations in a trajectory is 
Poisson(m_obs). The result is a number of trajectories, all starting at 
time 0, with different time spans, and with independently different numbers 
of observations within the time spans. Each trajectory follows a randomly 
selected response function with added N(mean, sd) error.
}

\examples{
set.seed(123)
data = gen_traj_data(n_id = 20, m_obs = 10, s_range = c(-50, -10),
              e_range = c(3*365, 10*365), reference = 100)
head(data)

}
