% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{gen_traj_data}
\alias{gen_traj_data}
\title{Data Generators}
\usage{
gen_traj_data(n_id, lambda_obs, first = c(-50, -10), last = c(50, 100),
  reference = 100, noise = c(0, abs(reference/20)), k = 3,
  plots = 0)
}
\arguments{
\item{n_id}{Number of trajectories}

\item{lambda_obs}{Poisson mean for number of observations in a trajectory
over 3. That is, the mean is lambda_obs + 3.}

\item{first}{Vector of length 2 giving first observation time range, 
c(min, max). First observation is uniformly distributed over this range.}

\item{last}{Vector of length 2 giving last observation time range, 
c(min, max). Last observation is uniformly distributed over this range.}

\item{reference}{A nominal response reference (for example, blood pressure
is near 100, which is the default)}

\item{noise}{Vector of length 2 giving the mean and standard deviation, 
c(mean, sd), of added N(mean, sd) noise.}

\item{plots}{Number of plots of randomly selected trajectories to check
generated data}
}
\description{
Collection of longitudinal responses with possibly varying lengths and 
varying numbers of observations. Support is [start . . . 0 . . . end] so that
all trajectories are aligned at 0. Usually, zero is the intervention date.
}
\section{Details}{

Generate longitudinal data for a response variable. Trajectories start
at time uniformly distributed in first and end at time uniformly 
distributed in last. Number of observations in a trajectory is 
Poisson(lambda_obs). The result is a number of trajectories, all starting at 
time 0, with different time spans, and with independently different numbers 
of observations within the time spans. Each trajectory follows a randomly 
selected response function with added N(mean, sd) error.
}

\section{Value}{

A data frame with one response per row and three columns: id (an integer in 
1:n_id), time (an iteger in 0 to last observation time), and response.
}

