% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{mse_g}
\alias{mse_g}
\alias{mae_g}
\alias{mme_g}
\alias{mxe_g}
\title{Various Loss functions used internally by clustra}
\usage{
mse_g(pred, id, response)

mae_g(pred, id, response)

mme_g(pred, id, response)

mxe_g(pred, id, response)
}
\arguments{
\item{pred}{Vector of predicted values.}

\item{id}{Integer vector of group assignments.}

\item{response}{Vector of response values.}
}
\value{
A numeric value. For mse_g(), returns the mean-squared error. For mae_g(),
returns mean absolute error. For mme_g(), returns median absolute error.
For mxe_g(), returns the
maximum absolute error.
}
\description{
Various Loss functions used internally by clustra
}
