% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{mse_g}
\alias{mse_g}
\alias{mxe_g}
\title{Loss functions}
\usage{
mse_g(pred, id, response)

mxe_g(pred, id, response)
}
\arguments{
\item{pred}{Vector of predicted values.}

\item{id}{Integer vector of group assignments.}

\item{response}{Vector of response values.}
}
\description{
\code{mse_g} Computes mean-squared loss for each group.
\code{mxe_g} is maximum loss.
}
