% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_sample}
\alias{plot_sample}
\title{Plots a sample of ids in a small mutiples layout}
\usage{
plot_sample(dat, layout = c(3, 3))
}
\arguments{
\item{dat}{A data frame with a few id trajectories to plot.}

\item{layout}{The small multiples layout as c(rows, columns).}
}
\value{
Invisibly returns the number of trajectories plotted.
}
\description{
Plots a sample of ids in a small mutiples layout
}
