% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{plot_tps}
\alias{plot_tps}
\title{Various functions to evaluate results of clustering, like RandIndex
computation and plots of spline fits.}
\usage{
plot_tps(dat, tps, file, points = FALSE)
}
\arguments{
\item{dat}{Data frame with variables *time*, *response*, *color*, *group*}

\item{tps}{A list, where each element is an output from \code{mgcv::bam}.}

\item{file}{Character string with the filename for pdf output.}

\item{points}{If true, plot all the data points. Not implemented yet.}
}
\description{
Plots spline fit (modify to use acutal bam results), possibly with data
TODO use actual tps fit rather than redo it with stat_smooth
}
\section{Details}{

}

