% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{pred_g}
\alias{pred_g}
\title{Function to predict for new data based on fitted tps of a group}
\usage{
pred_g(tps, data)
}
\arguments{
\item{tps}{Output structure of \code{\link[mgcv]{bam}}.}

\item{data}{See \code{\link{clustra}} description.}
}
\description{
Function to predict for new data based on fitted tps of a group
}
