% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{rand_plot}
\alias{rand_plot}
\title{Rand index matrix plot from Technometrics paper
Sorts replicates within cluster K
Assumes K starts from 2
Author: Wei-Chen Chen (wcsnow@gmail.com)}
\usage{
rand_plot(rand_pairs)
}
\arguments{
\item{rand_pairs}{A data frame with columns of cluster assignments}
}
\description{
Rand index matrix plot from Technometrics paper
Sorts replicates within cluster K
Assumes K starts from 2
Author: Wei-Chen Chen (wcsnow@gmail.com)
}
