% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{start_groups}
\alias{start_groups}
\title{Function to assign starting groups.}
\usage{
start_groups(data, k, nstart, nid, cores, maxdf, verbose = FALSE)
}
\arguments{
\item{data}{Data frame with response measurements, one per observation.
Column names are id, time, response, group. Note that id are already
sequential starting from 1. This affects expanding group numbers to ids.}

\item{k}{Number of clusters (groups).
"bestrand" is implemented.}

\item{maxdf}{Maximum degrees of freedom for trajectory spline smooths.}

\item{starts}{The number of random start values to consider.}

\item{snid}{The number of id's per cluster to sample in evaluating starts.}
}
\description{
Function to assign starting groups.
}
