% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{trajectories}
\alias{trajectories}
\title{Cluster longitudinal trajectories of a response variable.}
\usage{
trajectories(data, k, group, maxdf, iter, mccores, verbose = FALSE)
}
\arguments{
\item{data}{Data frame with response measurements, one per observation. Column
names are \code{id}, \code{time}, \code{response}, \code{group}. Note that
id must be already sequential starting from 1. This affects expanding group
numbers to ids.}

\item{k}{Number of clusters (groups)}

\item{group}{Vector of initial group numbers corresponding to ids.}

\item{maxdf}{Integer. Basis dimension of smooth term. See \code{\link[mgcv]{s}},
parameter k, in package \code{mgcv},}

\item{iter}{Integer. Maximum number of EM iterations.}

\item{mccores}{Integer number of cores to use by *mclapply* sections. Parallelization is 
over *k*, the number of clusters.}

\item{verbose}{Logical, whether to produce debug output.}
}
\description{
Trajectory means are splines fit to all ids in a cluster. Typically, this
function is called by \code{\link{clustra}}.
}
\author{
George Ostrouchov and David Gagnon
}
