% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{trajectories}
\alias{trajectories}
\title{Cluster longitudinal trajectories of a response variable.}
\usage{
trajectories(
  data,
  k,
  group,
  iter = clustra_env("clu$iter"),
  maxdf = clustra_env("clu$maxdf"),
  cores = clustra_env("cor"),
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame with response measurements, one per observation. Column
names are \code{id}, \code{time}, \code{response}, \code{group}. Note that
id must be already sequential starting from 1. This affects expanding group 
numbers to ids.}

\item{k}{Number of clusters (groups)}

\item{group}{Group numbers corresponding to sequential ids.}

\item{iter}{Maximum iterations in mgcv::bam (see .clustra_env)}

\item{maxdf}{Maximum degrees of freedom for tps in mgcv::bam (see .clustra_env)}

\item{cores}{List with cores allocation to various sections (see .clustra_env)}

\item{verbose}{Logical, whether to produce debug output.}
}
\description{
Trajectory means are splines fit to all ids in a cluster. Typically, this
function is called by \code{clustra()}.
}
\author{
George Ostrouchov and David Gagnon
}
