% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{trajectories}
\alias{trajectories}
\title{Cluster longitudinal trajectories of a response variable.}
\usage{
trajectories(
  data,
  k,
  group,
  fp,
  cores = c(e_mc = 1, m_mc = 1, nthreads = 1, blas = 1),
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame with response measurements, one per observation. Column
names are \code{id}, \code{time}, \code{response}, \code{group}. Note that
id must be already sequential starting from 1. This affects expanding group
numbers to ids.}

\item{k}{Number of clusters (groups)}

\item{group}{Group numbers corresponding to sequential ids.}

\item{fp}{List with additional fitting parameters: *maxdf* basis dimension of smooth
term (see \code{\link[mgcv]{s}}, parameter k, in package \code{mgcv}),
*iter* maximum number of EM iterations, *starts* number of trial random starts
from which the b  est is chosen, *idperstart* number of `id`s to sample for
evaluating random starts, *retry_max* number of restarts if iteration
encounters a cluster too small for \code{\link[mgcv]{bam}} fitting with the
given *maxdf*.}

\item{cores}{Vector of length 4. Core allocations to various sections of code: `e_mc` cores to use
by `mclapply` of e_mc (expectation over k), `m_mc` cores to use by 
`mclapply` of m_mc (maximization across k), `nthreads` threads to use by
\link[mgcv]{bam}, `blas` cores to use by OpenBLAS.}

\item{verbose}{Logical, whether to produce debug output.}
}
\description{
Trajectory means are splines fit to all ids in a cluster. Typically, this
function is called by \code{\link{clustra}}.
}
\author{
George Ostrouchov and David Gagnon
}
